refactor(offense): extract hardened offensive-runs capture/sign writer into shared module
Extract the signed-row capture/append/hash machinery out of offensive-idor-producer.js into mcp/lib/offensive-capture-writer.js so the upcoming reflected-XSS finder reuses the same security-reviewed write path (the PR-B move, applied to the writer). buildAndSignOffensiveRow is the parameterized buildSignedOffensiveRunRow (runIdPrefix/toolId/demonstratedSeverity/identityTag/stdoutContent/stderrContent); the helpers move verbatim; sha256OfFileFd stays internal; the producer drops the now-dead requires.

Behavior byte-identical: same 14-field row order, idor- run_id format, sign-last + hardened append. Full suite green (test:mcp 2605/0/2). Precursor to the reflected-XSS finder (PR_REFLECT_XSS_MVP_HANDOFF_PLAN).
